split untransform() into helper functions

+++ R/cvf.R
@@ -11,9 +11,13 @@
 #'
 #' @keywords internal
 cvf <- function(i, fold, type, cv_args, estimated_Sigma, ...) {
+
   # save the 'prep' object from the plmm_prep() in cv_plmm
   full_cv_prep <- cv_args$prep
+
+  # save outcome information -- will need to subset this into test and train sets
   y <- cv_args$y
+
   # make list to hold the data for this particular fold:
   fold_args <- list(std_X_details = list(),
                     fbm_flag = cv_args$fbm_flag,
@@ -69,21 +73,29 @@ cvf <- function(i, fold, type, cv_args, estimated_Sigma, ...) {
     if (sum(singular) >= 1) fold_args$penalty_factor[singular] <- Inf
 
   } else {
+
+    # subset training data (we will need 2 copies: a copy to pass into the
+    #   fitting function via the 'fold_args' list, and a copy to use
+    #   for prediction. The latter is 'train_X')
     fold_args$std_X <- train_X <- full_cv_prep$std_X[fold!=i, ,drop=FALSE]
+
     # Note: subsetting the data into test/train sets may cause low variance features
     #   to become constant features in the training data. The following lines address this issue
 
-    # re-scale data & check for singularity
-    fold_args$std_X <- ncvreg::std(fold_args$std_X) # notice: singular columns are *removed* here
-    fold_args$std_X_details$center <- attr(fold_args$std_X, "center")
-    fold_args$std_X_details$scale <- attr(fold_args$std_X, "scale")
-    fold_args$std_X_details$ns <- attr(fold_args$std_X, "nonsingular")
+    # re-standardize training data & check for singularity
+    std_info <- standardize_matrix(fold_args$std_X)
+    fold_args$std_X <- std_info$std_X
+    fold_args$std_X_details <- std_info$std_X_details
 
     # do not fit a model on these singular features!
-    fold_args$penalty_factor <- fold_args$penalty_factor[fold_args$std_X_details$ns]
+    singular <- fold_args$std_X_details$scale < 1e-3
+    if (sum(singular) >= 1) fold_args$penalty_factor[singular] <- Inf
 
   }
+  # re-center y
   fold_args$centered_y <- full_cv_prep$centered_y[fold!=i] |> scale(scale=FALSE) |> drop()
+
+  # subset outcome vector to include outcomes for training data only
   fold_args$y <- y[fold!=i]
 
   # extract test set --------------------------------------
@@ -97,8 +109,10 @@ cvf <- function(i, fold, type, cv_args, estimated_Sigma, ...) {
                                   backingpath = bigmemory::dir.name(full_cv_prep$std_X))
 
   } else {
-    test_X <- full_cv_prep$std_X[fold==i, fold_args$std_X_details$ns, drop=FALSE]
+    test_X <- full_cv_prep$std_X[fold==i, , drop=FALSE]
   }
+
+  # subset outcome for test set
   test_y <- y[fold==i]
 
   # decomposition for current fold ------------------------------
@@ -140,6 +154,8 @@ cvf <- function(i, fold, type, cv_args, estimated_Sigma, ...) {
                     convex = fold_args$convex,
                     dfmax = ncol(train_X) + 1)
 
+if (any(is.nan(fit.i$std_scale_beta))) browser()
+if (nrow(fit.i$std_scale_beta) - 1 != length(fold_args$std_X_details$ns)) browser()
   # get beta values back in original scale
   og_betas.i <- untransform(
     std_scale_beta = fit.i$std_scale_beta,
@@ -147,7 +163,6 @@ cvf <- function(i, fold, type, cv_args, estimated_Sigma, ...) {
     std_X_details = fold_args$std_X_details,
     fbm_flag = fold_args$fbm_flag,
     use_names = FALSE)
-
 
   if(type == "lp"){
     yhat <- predict_within_cv(fit = fit.i,
@@ -177,11 +192,13 @@ cvf <- function(i, fold, type, cv_args, estimated_Sigma, ...) {
   }
 
   # cleanup -----------------------------------------------------------------
-  # delete files created in cross-validation
-  list.files(path = bigmemory::dir.name(full_cv_prep$std_X),
-             pattern = paste0("fold",i),
-             full.names = TRUE) |> file.remove()
-  gc() # release the pointer
+  # delete files created in cross-validation, if data is filebacked
+  if (cv_args$fbm_flag) {
+    list.files(path = bigmemory::dir.name(full_cv_prep$std_X),
+               pattern = paste0("fold",i),
+               full.names = TRUE) |> file.remove()
+    gc() # release the pointer
+  }
 
   # return -----------------------------------------------------------------
   loss <- sapply(1:ncol(yhat), function(ll) plmm_loss(test_y, yhat[,ll]))
